Moving tqdm to misc
