--- src/amigo/tqdm.py
@@ -1,9 +1,0 @@
-# import tqdm appropriately
-from IPython import get_ipython
-
-if get_ipython() is not None:
-    # Running in Jupyter Notebook
-    from tqdm.notebook import tqdm
-else:
-    # Running in a script or other non-Jupyter environment
-    from tqdm import tqdm
